+++ lib/generators/react_on_rails/base_generator.rb
@@ -27,7 +27,7 @@ module ReactOnRails
           node_modules
 
           # Generated js bundles
-          /app/assets/webpack/*
+          /public/webpack/*
         DATA
 
         if dest_file_exists?(".gitignore")
@@ -35,26 +35,6 @@ module ReactOnRails
         else
           GeneratorMessages.add_error(setup_file_error(".gitignore", data))
         end
-      end
-
-      def update_application_js
-        data = <<-DATA.strip_heredoc
-          //= require webpack-bundle
-
-        DATA
-
-        app_js_path = "app/assets/javascripts/application.js"
-        found_app_js = dest_file_exists?(app_js_path) || dest_file_exists?("#{app_js_path}.coffee")
-        if found_app_js
-          prepend_to_file(found_app_js, data)
-        else
-          create_file(app_js_path, data)
-        end
-      end
-
-      def strip_application_js_of_double_blank_lines
-        application_js = File.join(destination_root, "app/assets/javascripts/application.js")
-        gsub_file(application_js, /^\n^\n/, "\n")
       end
 
       def create_react_directories
@@ -65,6 +45,8 @@ module ReactOnRails
       def copy_base_files
         base_path = "base/base/"
         base_files = %w(app/controllers/hello_world_controller.rb
+                        config/webpack/paths.yml
+                        config/webpack/development.server.yml
                         client/.babelrc
                         client/webpack.config.js
                         client/REACT_ON_RAILS_CLIENT_README.md)
@@ -73,14 +55,15 @@ module ReactOnRails
 
       def template_base_files
         base_path = "base/base/"
-        %w(config/initializers/react_on_rails.rb
+        %w(app/views/layouts/hello_world.html.erb
+           config/initializers/react_on_rails.rb
            Procfile.dev
            package.json
            client/package.json).each { |file| template("#{base_path}#{file}.tt", file) }
       end
 
       def add_base_gems_to_gemfile
-        append_to_file("Gemfile", "\ngem 'mini_racer', platforms: :ruby\n")
+        append_to_file("Gemfile", "\ngem 'mini_racer', platforms: :ruby\ngem 'webpacker_lite'\n")
       end
 
       ASSETS_RB_APPEND = <<-DATA.strip_heredoc
@@ -93,7 +76,7 @@ module ReactOnRails
 # Rails.application.config.assets.precompile += %w( server-bundle.js )
 
 # Add folder with webpack generated assets to assets.paths
-Rails.application.config.assets.paths << Rails.root.join("app", "assets", "webpack")
+Rails.application.config.assets.paths << Rails.root.join("public", "webpack", Rails.env)
       DATA
 
       def append_to_assets_initializer
@@ -137,6 +120,10 @@ Rails.application.config.assets.paths << Rails.root.join("app", "assets", "webpa
         message = <<-MSG.strip_heredoc
 
           What to do next:
+
+            - Include your webpack assets to your application layout.
+
+              <%= javascript_pack_tag 'main' %>
 
             - Ensure your bundle and yarn installs of dependencies are up to date.