+++ spec/dummy/client/server-rails-hot.js
@@ -16,16 +16,18 @@
 
 import webpack from 'webpack';
 import WebpackDevServer from 'webpack-dev-server';
-
+const { resolve } = require('path');
 import webpackConfig from './webpack.client.rails.hot.config';
 
-const hotRailsPort = process.env.HOT_RAILS_PORT || 3500;
+const webpackConfigLoader = require('react-on-rails/webpackConfigLoader');
+const configPath = resolve('..', 'config', 'webpack');
+const { devServer: devServerConfig, publicPath } = webpackConfigLoader(configPath);
 
 const compiler = webpack(webpackConfig);
 
 const devServer = new WebpackDevServer(compiler, {
-  contentBase: `http://lvh.me:${hotRailsPort}`,
-  publicPath: webpackConfig.output.publicPath,
+  contentBase: `http://lvh.me:${devServerConfig.port}`,
+  publicPath,
   hot: true,
   inline: true,
   historyApiFallback: true,
@@ -41,9 +43,9 @@ const devServer = new WebpackDevServer(compiler, {
   },
 });
 
-devServer.listen(hotRailsPort, 'localhost', err => {
+devServer.listen(devServerConfig.port, 'localhost', err => {
   if (err) console.error(err);
   console.log(
-    `=> 🔥  Webpack development server is running on port ${hotRailsPort}`
+    `=> 🔥  Webpack development server is running on port ${devServerConfig.port}`
   );
 });