+++ spec/dummy/client/webpack.client.base.config.js
@@ -2,29 +2,35 @@
 // webpack.client.rails.hot.config and webpack.client.rails.build.config.
 
 const webpack = require('webpack');
-const path = require('path');
+const { resolve, join } = require('path');
 const webpackCommon = require('./webpack.common');
 const { assetLoaderRules } = webpackCommon;
+
+const ManifestPlugin = require('webpack-manifest-plugin');
+const webpackConfigLoader = require('react-on-rails/webpackConfigLoader');
+const configPath = resolve('..', 'config', 'webpack');
+const { paths, publicPath } = webpackConfigLoader(configPath);
 
 const devBuild = process.env.NODE_ENV !== 'production';
 
 module.exports = {
 
   // the project dir
-  context: __dirname,
+  context: resolve(__dirname),
   entry: {
     // This will contain the app entry points defined by
     // webpack.client.rails.hot.config and webpack.client.rails.build.config
-    app: [
+    'app-bundle': [
       './app/startup/clientRegistration',
     ],
   },
   resolve: {
     extensions: ['.js', '.jsx'],
     alias: {
-      images: path.join(process.cwd(), 'app', 'assets', 'images'),
+      images: join(process.cwd(), 'app', 'assets', 'images'),
     },
   },
+
 
   plugins: [
     new webpack.EnvironmentPlugin({ NODE_ENV: 'development' }),
@@ -34,12 +40,13 @@ module.exports = {
 
     // https://webpack.js.org/guides/code-splitting-libraries/#implicit-common-vendor-chunk
     new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
+      name: 'vendor-bundle',
       minChunks(module) {
         // this assumes your vendor imports exist in the node_modules directory
         return module.context && module.context.indexOf('node_modules') !== -1;
       },
     }),
+    new ManifestPlugin({ fileName: paths.manifest, publicPath, writeToFileEmit: true }),
   ],
 
   module: {