+++ spec/dummy/client/webpack.client.rails.build.config.js
@@ -6,15 +6,18 @@
 
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const config = require('./webpack.client.base.config');
-const path = require('path');
+const { resolve } = require('path');
+const webpackConfigLoader = require('react-on-rails/webpackConfigLoader');
+const configPath = resolve('..', 'config', 'webpack');
+const { paths, publicPath } = webpackConfigLoader(configPath);
 
 const devBuild = process.env.NODE_ENV !== 'production';
 
 config.output = {
-  filename: '[name]-bundle.js',
-  path: path.resolve(__dirname, '../app/assets/webpack'),
-  publicPath: '/assets/',
-};
+  filename: '[name].js',
+  path: resolve('..', paths.output, paths.assets),
+  publicPath,
+},
 
 // See webpack.client.base.config for adding modules common to both the webpack dev server and rails
 
@@ -95,7 +98,7 @@ config.module.rules.push(
 
 config.plugins.push(
   new ExtractTextPlugin({
-    filename: '[name]-bundle.css',
+    filename: '[name].css',
     allChunks: true
   })
 );